+++ Libraries/LibGfx/ImmutableBitmap.h
@@ -10,37 +10,12 @@
 #include <AK/AtomicRefCounted.h>
 #include <AK/Forward.h>
 #include <AK/NonnullRefPtr.h>
-#include <LibGfx/BitmapExportResult.h>
 #include <LibGfx/Color.h>
 #include <LibGfx/ColorSpace.h>
 #include <LibGfx/Forward.h>
 #include <LibGfx/Rect.h>
 
 namespace Gfx {
-
-#define ENUMERATE_EXPORT_FORMATS(X) \
-    X(Gray8)                        \
-    X(Alpha8)                       \
-    X(RGB565)                       \
-    X(RGBA5551)                     \
-    X(RGBA4444)                     \
-    X(RGB888)                       \
-    X(RGBA8888)
-
-enum class ExportFormat : u8 {
-#define ENUMERATE_EXPORT_FORMAT(format) format,
-    ENUMERATE_EXPORT_FORMATS(ENUMERATE_EXPORT_FORMAT)
-#undef ENUMERATE_EXPORT_FORMAT
-};
-
-[[nodiscard]] StringView export_format_name(ExportFormat);
-
-struct ExportFlags {
-    enum : u8 {
-        PremultiplyAlpha = 1 << 0,
-        FlipY = 1 << 1,
-    };
-};
 
 class ImmutableBitmap final : public AtomicRefCounted<ImmutableBitmap> {
 public:
@@ -57,8 +32,6 @@ public:
 
     AlphaType alpha_type() const;
     ColorSpace const& color_space() const;
-
-    [[nodiscard]] ErrorOr<BitmapExportResult> export_to_byte_buffer(ExportFormat format, int flags, Optional<int> target_width, Optional<int> target_height) const;
 
     Color get_pixel(int x, int y) const;
